Use GoMiler struct and create gitlab.go for GitLab specific functions
The use of milestoneAPI for both GitLab and GitHub APIs worked, but only by coincidence. Using a generic struct was needed to better handle the differences between APIs.
